Refactor in a lib and add test
